Adding validation for the undefined value in javascript.
This change will help to remove the display of the value undefined in grid.
Following are the areas that are affected:
ContactListpopup of Lexmark Services Portal and Lexmark OPS Portal

This issue occurred as the value was blank and $("").val() returned undefined.
So, had to check the value returned by jQuery is undefined or not.

HPQC defect: 21472.