Retrieving all the serviceRequestLocale values based on Flag
This change has to be done because of the other performance issue.
Pull request "Changes for loading Service Request Detail Page loading time".
For the above change the ServiceRequestLocale values were not getting populated in
the Administration tab.
Now, on the basis of the flag in SRAdministrationListContract the values will be populated.

Impacted Areas: Administration Tab LexmarkServicesPortal , LexmarkOPSPortal.

Admin will be able to view the serviceRequestLocale values in portal correctly.