Removed Hardcoded validation Field and added File Cache version
Following are the changes that are done.
1) Added file cache versions to the js include.
2) The hardcoded validation 'formatValidation[13][1].indexOf("Consumables")' was wrong
as it was always returning true. As a result else part was not getting executed.
So, the checking should be on the basis of that particular key. Not the 13th field.
3) removed the Empty entry from the hwDebriefDeinstallCloseOut.jsp. As it is not required
and it was generating an error.

Impacted Areas: Partner Portal Changemanagement requests update.

Result: The ip address validation which was not working because of the hardcoded value, is
working properly now.

Related CR:  CR #CHG0010145 - Service Now