Partners are not able to close out a claim.
The Partner closes out and debrief a claim from portal but when they click continue
nothing happens. At the time of investigating we found there is a javascript error where few
variables are not reachable. We fixed the issue putting the variables in such a way so that it can be reachable by the working code.
This changes is not associated with any other functional module.