Updating the Price Password in QA environment.
SAP have reset and updated the passwords to "sjEbEL123" from "Siebel07".
Keeping the old password might result in hitting the SAP urls with the wrong password which might lock down the SAP price servers.
Hence changing the passwords on Portal built property files and as well as in individual project property files.