Mapping added for Kaiser Public URL.
A/c to the requirement, User should be able to hit a public URL to view the Ariba catalog without logging into Ariba.
https://portal-qa.lexmark.com/web/global-services-punchout/home-kaiser
This URL will hide the sensitive informations like Price and will disable the capability to order Items from the portal. The sole purpose of the URl is to expose the ariba catalog to the user.
Hence this URL has to be mapped in the liferay-hook.xml for redirection the flow control to the respective servlet in the application.