SR icons are displayed on device tile, but Request history returns no results.
Portal used to show the open MADC SR on the MAP left nav if the SR is <90 days,  But WM used the logic to show all the open SRs on the map.
Due to this discrepancy , portal has to remove the date range criteria for loading the open MADC SRs .
Removing the date range filter criteria will fetch all the open MADC SRs from siebel. This implementation is in sync with WM implementation in case of open MADC SRs.

Defect # 21257