Added the grouping of Model Type Family for Options and Warranties.
This change is for grouping multiple options and warranties based on
model type family.
JsonUtil - The change is to prepare the json with the grouping of the model family
bundle.js - Parsing of the json data and setting the accessories.
cart.js - Once the user adds to cart we have to add all the item in the group to cart.
bundle.jsp - The structure of the table is moved to bundle list.