Changes done for the Cache Entry.
The cache  retrieveOptionsAndWarranties and retrieveBundleList is used in Lexmark services Punchout .
These cache the data retrieved from the aMind call. As the aMind jar is commonly used in Services , OPS and Partner portal , hence the cache entry must be present in all the three projects.
This checkin contains the cache xml files with the entry mentioned above.