Removing SoldTo number in Catalog Order.
The changes are made to remove the sold to number from the Payment type call , Catalog List call , Part and Model list call and in the Tax call.
This will facilitate the user to view the list of data on the basis of the contract selected and irrespective of the sold to.

The removal of Sold to number from the aMind contracts have been done in OrderSuppliesCatalogController.java both in Customer Portal and OPS Portal.

Defect # 22769