Theme Content-Footer and Cancel Button UI Changes
This change is for the Footer and Content section in Lexmark Punchout Portal.
Its look and feel will be same as that of shop.lexmark.com/kaiser.

The look and feel of the cancel button in the Punchout portal is changed
according to the new style (same as that of shop.lexmark.com/kaiser). The
changes are for its padding, border, background-color, font-size and font-color.