Add Delete button during SR update
This change is required to provide an option for the partners to delete
the selected part at the time of SR update in Partner Portal. Earlier there
was no delete button which was causing the partner to Go Back if they selects
any part by mistake just to discard the selected part.

We have created one Delete Part button on every part row items so that
the partners can delete the wrong part without going back to the previous page.

This is related to the Service now Change #CHG0001321