PDF doesn't have the Technician Instruction section
While the partners or FSEs are printing the Service Request, the Technician
Instruction filed is missing. So in order to display the Technician Instruction
in the PDF code change is required so that it capture the mentioned information.
This change only applicable to Partner Portal Service Request PDF generation.

This requirement is as per the AMS CR #CHG0009981.