Modification of the GLS-Italy Carrier link
This change is related to the modification of the GLS-Italy Carrier site.
The GLS-Italy carrier link has been modified by the vendor. So currenly the
tracking link in Customer Portal is not getting redirected to the correct tracking page. In order
to make it working we need to use the modified link as provided.

This requirement is as per the AMS CR #CHG0015508