Changes for visiblility of createRequest Menu in Fleet Manager of OPS Portal
This change is for making CreateRequest Menu visible in Floor Level Summary in Fleet Management Tab of OPS Portal.
Issue: On clicking Create Request link the Create Request Menu was not appearing in the browser window. User had to scroll to see the menu.
The above changes address this issue.
HPQC CR Number - 14645