Stop Duplicate Order for Consumable Install activities in Partner Portal
In current scenario, the Partners & FSEs are able to place duplicate orders although the
order is placed. This change is required in Partner portal to restrict the partners &
FSEs to place duplicate Orders. This change is required in Partner Portal in Service Request
Update call.

The related Service Now ticket number is CHG0009901/INC0166687.