Customer Request tab in Partner Portal Performance Issue.
This is related to the Partner Portal Customer request tab performance issue where the partners has to wait for the grid to load after selecting the account in the pop up. The fix is to stop the duplicate call for getting entitlement flag as that call is consuming around 3 minutes to load. This change is exclusively for partner Portal

The related incident/change number is INC0220223/CHG0014471