Restrict Document Based on the Partner Country Set up.
This is related to the visibility of the documents in the Documents Library tab in Partner portal. In Current scenario there is no check to restrict the document based on the Partner Country Set Up. This change will enable the functionality of document visibility based on the Partner Country.

The related incident/CR # is INC0191549/CHG0011990